requirements.txt
asgiref==3.9.1
Django==5.2.5
PyMySQL==1.1.1
python-dotenv==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
wheel==0.45.1
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-crispy-forms==2.3

# Phase 1 Dependencies
Pillow==10.1.0
redis==5.0.1
celery==5.3.4
django-extensions==3.2.3

# Phase 2 Dependencies (for future use)
# elasticsearch==8.11.0
# django-elasticsearch-dsl==8.0
# geopy==2.4.1
# stripe==7.9.0
# requests==2.31.0